==> GSE88884TRIVAL/03_PCA_analysis_haoLiu.Rmd <==
---
title: "Untitled"
author: '520595273'
date: "2025-04-04"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(dplyr)
library(tidyr)
library(ggplot2)
library(GEOquery)
library(limma)
```

```{r}
expr_df <- readRDS("expr_df_matched.rds")  
pheno_df <- readRDS("pheno_clean.rds")
```

```{r}
colnames(expr_df)
head(expr_df)
```

```{r}
colnames(pheno_df)
head(pheno_df)
```

```{r}
pheno_df$SampleID <- rownames(pheno_df)
all(colnames(expr_df) %in% pheno_df$SampleID)
```

```{r}
gene_variance <- apply(expr_df, 1, var)
top_genes <- names(sort(gene_variance, decreasing = TRUE))[1:200]
expr_df_top <- expr_df[top_genes, ]
```


```{r}
expr_for_pca <- t(expr_df_top)
pca_res <- prcomp(expr_for_pca, center = TRUE, scale. = TRUE)
pca_df <- as.data.frame(pca_res$x)
pca_df$SampleID <- rownames(pca_df)
merged_df <- merge(pca_df, pheno_df, by = "SampleID")
```



```{r}
ggplot(merged_df, aes(x = PC1, y = PC2, color = c3_status)) +
  geom_point(size = 2) +
  facet_wrap(~ c4_status) +  
  theme_bw() +
  labs(
    title = "PCA by C3 status (color) & C4 status (facet)",
    x = "PC1", y = "PC2",
    color = "C3 status"
  )
```

```{r}
ggplot(merged_df, aes(x = PC1, y = PC2, 
                      color = antidsdna_at_baseline)) +
  geom_point(size = 3) +
  theme_light() +
  labs(
    title = "PCA by anti-dsDNA (color) & sex (shape)",
    x = "PC1", y = "PC2",
    color = "anti-dsDNA",
    shape = "Sex"
  )
```

```{r}
merged_df$c3c4_combo <- paste(merged_df$c3_status, merged_df$c4_status, sep = "_")

ggplot(merged_df, aes(x = PC1, y = PC2, color = c3c4_combo)) +
  geom_point(size = 3) +
  theme_bw() +
  labs(
    title = "PCA by combined C3 and C4 status",
    x = "PC1", y = "PC2",
    color = "C3_C4"
  )
```

```{r}
ggplot(merged_df, aes(x = PC1, y = PC2, color = sledai_at_baseline)) +
  geom_point(size = 2) +
  theme_bw() +
  labs(
    title = "PCA by SLEDAI value",
    x = "PC1", y = "PC2",
    color = "SLEDAI value"
  )
```



```{r}
sdev <- pca_res$sdev
var_explained <- sdev^2 / sum(sdev^2)

plot(var_explained[1:20], type = "b", 
     xlab = "Principal Component", ylab = "Proportion of Variance Explained",
     main = "Scree Plot")

```


```{r}
library(conflicted)      
conflict_prefer("select", "dplyr")  

library(dplyr)
library(ggcorrplot)

num_vars <- pheno_df |>
  dplyr::select(where(is.numeric), -time)

cor_mat <- cor(num_vars,
               use    = "pairwise.complete.obs",
               method = "spearman")

ggcorrplot(cor_mat,
           type  = "lower",
           lab   = TRUE,
           tl.cex = 8,
           title = "Spearman ρ among numeric phenotypes")

```


```{r}
library(ggplot2)
library(ggrepel)
library(conflicted)            
conflict_prefer("as.data.frame", "base")  
 
loadings <- base::as.data.frame(pca_res$rotation[, 1:2])
loadings$var <- rownames(loadings)

loadings <- loadings |>
  dplyr::mutate(across(c(PC1, PC2), ~ .x * 3))

ggplot(merged_df, aes(PC1, PC2)) +
  geom_point(aes(color = antidsdna_at_baseline,
                 shape = c3_status), size = 2) +
  geom_segment(data = loadings,
               aes(x = 0, y = 0, xend = PC1, yend = PC2),
               arrow = arrow(length = unit(0.2, "cm")),
               linewidth = .6, color = "grey30") +
  geom_text_repel(data = loadings,
                  aes(x = PC1, y = PC2, label = var),
                  size = 3.5) +
  scale_color_viridis_c(option = "plasma") +
  labs(title = "PCA biplot: samples + variable loadings",
       x = "PC1", y = "PC2",
       color = "anti‑dsDNA", shape = "C3 status") +
  theme_minimal()


```
 
 
```{r}
library(ggplot2)

ve <- pca_res$sdev^2 / sum(pca_res$sdev^2) * 100

ggplot(merged_df, aes(PC1, PC2, colour = sledai_at_baseline)) +
  stat_ellipse(type = "t", level = 0.95, colour = "grey50") +   
  geom_point(size = 2) +
  geom_hline(yintercept = 0, linetype = "dashed") +
  geom_vline(xintercept = 0, linetype = "dashed") +
  labs(title = "PCA of Gene Expression Data",
       x = sprintf("PC1 (%.1f%% Variance Explained)", ve[1]),
       y = sprintf("PC2 (%.1f%% Variance Explained)", ve[2]),
       colour = "SLEDAI") +
  scale_colour_viridis_c(option = "plasma") + 
  coord_equal() +
  theme_minimal()


```
